+++ CHANGELOG.md
@@ -7,7 +7,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-...
+### Fixed
+- Reader conditional macros in namespace forms are correctly handled at the top
+  level and inside `:require` type forms. Some other cases (such as conditionals
+  inside a libspec) may still cause errors.
+  [#5](//github.com/greglook/cljfmt/issues/5)
 
 ## [0.8.3] - 2019-10-08